Moved tables with async fetch in their own component
