+++ src/views/ClientList.tsx
@@ -1,101 +1,16 @@
 import React, { FunctionComponent, useContext } from 'react'
-import { Client, ClientKind, ActionProps, Party, MappedRouteConfig } from '../../types'
+import { ClientKind } from '../../types'
 import { StyledIntro } from '../components/Shared/StyledIntro'
-import { TableWithLoader } from '../components/Shared/TableWithLoader'
 import { TempFilters } from '../components/TempFilters'
-import { useAsyncFetch } from '../hooks/useAsyncFetch'
 import { isAdmin } from '../lib/auth-utils'
 import { PartyContext } from '../lib/context'
-import { buildDynamicPath } from '../lib/router-utils'
 import { StyledButton } from '../components/Shared/StyledButton'
-import { StyledTableRow } from '../components/Shared/StyledTableRow'
-import { ActionMenu } from '../components/Shared/ActionMenu'
-import { useHistory } from 'react-router-dom'
-import { axiosErrorToError } from '../lib/error-utils'
 import { useRoute } from '../hooks/useRoute'
-import { useFeedback } from '../hooks/useFeedback'
 import { PageTopFilters } from '../components/Shared/PageTopFilters'
-import { Box } from '@mui/material'
+import { AsyncTableClient } from '../components/Shared/AsyncTableClient'
 
 type ClientListProps = {
   clientKind?: ClientKind
-}
-
-type AsyncTableProps = {
-  clientKind: ClientKind
-  party: Party | null
-  routes: Record<string, MappedRouteConfig>
-}
-
-const AsyncTable = ({ clientKind, party, routes }: AsyncTableProps) => {
-  const { runAction, forceRerenderCounter } = useFeedback()
-
-  const history = useHistory()
-  const editPath =
-    clientKind === 'CONSUMER'
-      ? routes.SUBSCRIBE_CLIENT_EDIT.PATH
-      : routes.SUBSCRIBE_INTEROP_M2M_CLIENT_EDIT.PATH
-
-  const { data, loadingText, error } = useAsyncFetch<{ clients: Array<Client> }, Array<Client>>(
-    {
-      path: { endpoint: 'CLIENT_GET_LIST' },
-      config: { params: { kind: clientKind, consumerId: party?.id } },
-    },
-    {
-      mapFn: (data) => data.clients,
-      loaderType: 'contextual',
-      loadingTextLabel: 'Stiamo caricando i client',
-      useEffectDeps: [forceRerenderCounter],
-    }
-  )
-
-  /*
-   * List of possible actions for the user to perform
-   */
-  const wrapDelete = (clientId: string) => async () => {
-    await runAction(
-      { path: { endpoint: 'CLIENT_DELETE', endpointParams: { clientId } } },
-      { showConfirmDialog: true }
-    )
-  }
-  /*
-   * End list of actions
-   */
-
-  const getAvailableActions = (client: Client): Array<ActionProps> => {
-    return [{ onClick: wrapDelete(client.id), label: 'Elimina' }]
-  }
-
-  const headData = ['Nome client', '']
-
-  return (
-    <TableWithLoader
-      loadingText={loadingText}
-      headData={headData}
-      noDataLabel="Non ci sono client disponibili"
-      error={axiosErrorToError(error)}
-    >
-      {data &&
-        Boolean(data.length > 0) &&
-        data.map((item, i) => (
-          <StyledTableRow key={i} cellData={[{ label: item.name }]}>
-            <StyledButton
-              variant="outlined"
-              size="small"
-              onClick={() => {
-                history.push(buildDynamicPath(editPath, { clientId: item.id }))
-              }}
-            >
-              Ispeziona
-            </StyledButton>
-
-            <Box component="span" sx={{ ml: 2, display: 'inline-block' }}>
-              <ActionMenu actions={getAvailableActions(item)} />
-            </Box>
-          </StyledTableRow>
-        ))}
-    </TableWithLoader>
-  )
 }
 
 export const ClientList: FunctionComponent<ClientListProps> = ({ clientKind = 'CONSUMER' }) => {
@@ -128,7 +43,7 @@ export const ClientList: FunctionComponent<ClientListProps> = ({ clientKind = 'C
         )}
       </PageTopFilters>
 
-      <AsyncTable clientKind={clientKind} party={party} routes={routes} />
+      <AsyncTableClient clientKind={clientKind} />
     </React.Fragment>
   )
 }
