+++ src/views/EServiceManage.tsx
@@ -1,19 +1,12 @@
-import React, { useContext } from 'react'
+import React from 'react'
 import { StyledIntro } from '../components/Shared/StyledIntro'
-import { DialogContext } from '../lib/context'
 import { StyledButton } from '../components/Shared/StyledButton'
 import { Tab } from '@mui/material'
 import { TabList, TabContext, TabPanel } from '@mui/lab'
 import { useActiveTab } from '../hooks/useActiveTab'
 import { EServiceContentInfo } from '../components/Shared/EServiceContentInfo'
-import { ActionProps, DecoratedPurpose, EServiceReadType, Purpose } from '../../types'
-import { TableWithLoader } from '../components/Shared/TableWithLoader'
-import { formatThousands } from '../lib/format-utils'
-import { StyledTableRow } from '../components/Shared/StyledTableRow'
-import { decoratePurposeWithMostRecentVersion } from '../lib/purpose'
-import { formatDateString } from '../lib/format-utils'
-import { RunAction, useFeedback } from '../hooks/useFeedback'
-import { ActionMenu } from '../components/Shared/ActionMenu'
+import { EServiceReadType } from '../../types'
+import { useFeedback } from '../hooks/useFeedback'
 import { useLocation } from 'react-router-dom'
 import {
   decorateEServiceWithActiveDescriptor,
@@ -23,109 +16,7 @@ import { useAsyncFetch } from '../hooks/useAsyncFetch'
 import { NotFound } from './NotFound'
 import { useRoute } from '../hooks/useRoute'
 import { PageBottomActions } from '../components/Shared/PageBottomActions'
-
-type AsyncTableProps = {
-  forceRerenderCounter: number
-  runAction: RunAction
-  eserviceId?: string
-}
-
-const AsyncTable = ({ forceRerenderCounter, runAction, eserviceId }: AsyncTableProps) => {
-  const { setDialog } = useContext(DialogContext)
-
-  const { data: purposeData /* , error */ } = useAsyncFetch<
-    { purposes: Array<Purpose> },
-    Array<DecoratedPurpose>
-  >(
-    {
-      path: { endpoint: 'PURPOSE_GET_LIST' },
-      config: { params: { states: 'WAITING_FOR_APPROVAL', eserviceId } },
-    },
-    {
-      mapFn: (data) => data.purposes.map(decoratePurposeWithMostRecentVersion),
-      loadingTextLabel: 'Stiamo caricando le finalità in attesa',
-      useEffectDeps: [forceRerenderCounter],
-    }
-  )
-
-  /*
-   * List of possible actions for the user to perform
-   */
-  const wrapUpdatePurposeExpectedApprovalDate =
-    (purposeId: string, versionId: string, approvalDate?: string) => () => {
-      setDialog({
-        type: 'setPurposeExpectedApprovalDate',
-        purposeId,
-        versionId,
-        approvalDate,
-        runAction,
-      })
-    }
-
-  const wrapActivatePurpose = (purposeId: string, versionId: string) => async () => {
-    await runAction(
-      {
-        path: {
-          endpoint: 'PURPOSE_VERSION_ACTIVATE',
-          endpointParams: { purposeId, versionId },
-        },
-      },
-      { showConfirmDialog: true }
-    )
-  }
-  /*
-   * End list of actions
-   */
-
-  const getAvailableActions = (item: DecoratedPurpose): Array<ActionProps> => {
-    const actions = [
-      {
-        onClick: wrapUpdatePurposeExpectedApprovalDate(
-          item.id,
-          item.mostRecentVersion.id,
-          item.mostRecentVersion.expectedApprovalDate
-        ),
-        label: 'Aggiorna data di completamento',
-      },
-      {
-        onClick: wrapActivatePurpose(item.id, item.mostRecentVersion.id),
-        label: 'Attiva',
-      },
-    ]
-    return actions
-  }
-
-  const headData = ['Nome finalità', 'Stima di carico', 'Data di completamento', '']
-
-  return (
-    <TableWithLoader
-      loadingText={null}
-      headData={headData}
-      noDataLabel="Nessuna finalità da evadere"
-    >
-      {purposeData &&
-        Boolean(purposeData.length > 0) &&
-        purposeData.map((item, i) => {
-          return (
-            <StyledTableRow
-              key={i}
-              cellData={[
-                { label: item.title },
-                { label: formatThousands(item.mostRecentVersion.dailyCalls) },
-                {
-                  label: item.mostRecentVersion.expectedApprovalDate
-                    ? formatDateString(item.mostRecentVersion.expectedApprovalDate)
-                    : 'In attesa di presa in carico',
-                },
-              ]}
-            >
-              <ActionMenu actions={getAvailableActions(item)} />
-            </StyledTableRow>
-          )
-        })}
-    </TableWithLoader>
-  )
-}
+import { AsyncTablePurposeInEService } from '../components/Shared/AsyncTablePurpose'
 
 export function EServiceManage() {
   const { routes } = useRoute()
@@ -177,7 +68,7 @@ export function EServiceManage() {
           </React.Fragment>
         </TabPanel>
         <TabPanel value="purposeAwaitingApproval">
-          <AsyncTable
+          <AsyncTablePurposeInEService
             forceRerenderCounter={forceRerenderCounter}
             runAction={runAction}
             eserviceId={eserviceId}
